+++ hack/runtime-migrator/internal/config/config.go
@@ -1,4 +1,4 @@
-package internal
+package config
 
 import (
 	"flag"
